+++ src/OpenTelemetry.Exporter.Geneva/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+* Support for serializing `ActivityEvent`s.
 
 * Add ResourceFieldNames to filter resource attributes to send to Geneva
   ([#3552](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3552))